requirements.txt
config-parser==0.0.1
Django==1.11.2
django-ckeditor==5.9.0
django-filter==1.0.4
django-filters==0.2.1
django-js-asset==1.2.2
django-redis==4.11.0
djangorestframework==3.6.3
mysqlclient==1.4.6
olefile==0.44
Pillow==4.1.1
pkg-resources==0.0.0
pytz==2017.2
PyYAML==5.3
redis==3.4.1
